Adopt REUSE machine-readable licensing standard
The copyright headers were initially added by myself using the GPLv3 only, but
the original MDA plugins were released under both the GPLv2+ and MIT licenses.
My intent was always to retain the original license in this port (as described
in the README), so this commit explicitly licenses everything as GPLv2+ and
MIT for simplicity.